creacion del proceso compras

+++ database/migrations/2023_08_18_031339_create_compra_productos_table.php
@@ -8,17 +8,20 @@ class CreateCompraProductosTable extends Migration
 {
     public function up()
     {
+        if(Schema::hasTable('compra_productos')) return;       //add this line to migration file
         Schema::create('compra_productos', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('id_compra'); // Columna que referencia a la compra
-            $table->unsignedBigInteger('id_producto'); // Columna que referencia al producto
+            $table->unsignedBigInteger('id_compra');
+            $table->unsignedBigInteger('id_producto');
             $table->integer('cantidad');
-            $table->decimal('precio_unitario', 10, 2);
+            $table->decimal('precio_unitario', 10, 2); // Agregar esta línea
             $table->timestamps();
 
             $table->foreign('id_compra')->references('id_compra')->on('compras')->onDelete('cascade');
             $table->foreign('id_producto')->references('id_producto')->on('productos')->onDelete('cascade');
         });
+    
+        
     }
 
     public function down()
